pyproject.toml
[tool.poetry]
name = "balance"
version = "1.0a"
description = "API for User Balances."
authors = ["likecompany"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
likeinterface = "^1.1.0.1a"
corecrud = "^0.1.4.1"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
colorlog = "^6.7.0"
stringcase = "^1.2.0"
starlette-admin = "^0.10.0"
pydantic = {extras = ["email"], version = "^2.1.1"}
pydantic-settings = "^2.0.2"
fastapi = "^0.100.0"
uvicorn = "^0.23.1"
sqlalchemy = "^2.0.7"
alembic = "^1.10.2"
asyncpg = "^0.28.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.2.2"
black = "^23.3.0"
ruff = "^0.0.278"
isort = "^5.12.0"
mypy = "^1.2.0"

[tool.poetry.group.tests.dependencies]
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.0.0"
httpx = "^0.24.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = [
    "tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
omit = [
    "*__init__.py",
    "*orm",
    "*exc.py",
    "*app.py",
]

[tool.coverage.report]
exclude_lines = [
    "if TYPE_CHECKING:",
    "pragma: no cover",
]

[tool.coverage.html]
directory = "htmlcov"


[tool.ruff]
line-length = 99
select = [
    "C4",
    "E",
    "F",
    "T10",
    "T20",
    "Q",
    "RET",
]
ignore = [
    "E501",
]
exclude = [
    ".git",
    ".venv",
    "alembic",
    "logs",
]
force-exclude = true

[tool.black]
line-length = 99
target-version = ['py38', 'py39', 'py310', 'py311']
exclude = '''
(
    \.git
    \.ruff_cache
    \.venv
  | alembic
  | logs
)
'''

[tool.isort]
profile = "black"
src_paths = [
    "app",
    "tests",
]
skip = [
    "alembic",
]

[tool.mypy]
plugins = [
    "pydantic.mypy",
]
disable_error_code = [
    "import",
]
show_error_codes = true
show_error_context = true
namespace_packages = true

[[tool.mypy.overrides]]
module = [
    "passlib",
]
ignore_missing_imports = true
disallow_untyped_defs = true
